src/gui/requirements.txt
PyQt5 (pip3)
pyttsx3 (pip3)
espeak (sudo apt install espeak)
rospy
pyyaml(pip3)
datetime(pip3)
